SCORSRVR-20. Fix the leaky classloader issues.
Avro, Hadoop etc have handles to the classloader that prevents the webapp
classloader from properly unloading when a given model is undeployed. This is
problematic in a world where you have models coming and going which would lead
to an eventual exhaustion of the permgen space. Clearing out some of these
references or in the extreme case, disabling Avro Class caching (which
empirically has been shown to not really have a material impaact because it
seems that the ClassLoader itself caches stuff) removes these potential
opportunities for ClassLoader leaks.

Reviewer: robert, ajprax
Review: https://review.kiji.org/r/1791/